extracting more indexing logic
